Add yarn.lock file and update .gitignore
The yarn.lock file has been initialized and populated with dependencies for the project. The .gitignore file has also been updated to no longer ignore the yarn.lock file so that it can be tracked by version control. This will help ensure everyone is using the same versions of the dependencies.