Update GameStorage.sol with strain data changes
In this commit, the GameStorage.sol file was updated. The most significant change is the replacement of the strainTotalSupply mapping with a new strainBurned mapping. Convenience methods to calculate total supply and total left of each strain were also introduced. Indentation inconsistencies in createPlant and createItem functions were corrected.

Remove redundant and commented out code

Removed a redundant check for strain supply limit and commented out code from the GameLogic.sol file. This is part of ongoing efforts to clean up and streamline the codebase.