Add OpenZeppelin contracts remapping and comment out ERC2771Forwarder code
A remapping for '@openzeppelin/contracts-upgradeable/' was added to the 'foundry.toml' file. Several lines related to the ERC2771ForwarderUpgradeable contract have been commented out in 'GameLogic.sol' and 'PixotchiRouter.sol' files. These changes suggest we're reorganizing the code to no longer utilise ERC2771Forwarder or to prepare for a different implementation.