Refactor isApproved modifier and initialization
In this update, the `isApproved` modifier in GardenLogic.sol was removed and replaced with a function `isApprovedFn` in the IGame interface. Furthermore, the initialization logic in ConfigLogic.sol and GameLogic.sol was updated. Additional minor changes include adjusting the build script in package.json and commenting out some imports in GameLogic.sol.