Add GameLogic, ConfigLogic and other extensions to the deployment script
A JSON deployment script has been added, with details for multiple extensions including ConfigLogic and GameLogic. This script includes metadata, such as the name and IPFS metadata URI, as well as function selectors and signatures for each extension. It also details the bytecode for each extension's implementation. This will help to manage and streamline the deployment process for these extensions.