Add initial time of day (TOD) parameter to strain
In this update, we've made changes to the strain and game logic. We've added an initial time of day (TOD) parameter to the strain structure, which allows admins to set a different initial TOD for each strain. We've also introduced public visibility to the isApprovedFn function and added a redeem function.