DEPOYMENT Simplify get shop items method and add getPlantName function
The get purchased shop items method has been simplified to return a single static shop item. This has minimized the complexity and made the result predictable for testing purposes. Additionally, a 'getPlantName' function was added to the 'GameLogic' and 'IPixotchi' contracts to retrieve the name of a plant given an ID.