Change visibility of shopIsEffectOngoing and rearrange purchase logic
The function shopIsEffectOngoing is now public instead of internal, allowing other contracts to see if an effect is still ongoing for a specific NFT and item. The total consumed count for an item is now incremented after the payment validation process and not before, ensuring consistency in item consumption tracking. Furthermore, the signature for this method has been added to the IPixotchi interface to standardize its usage across different contracts.