Update package versions in package.json
The version numbers for `@openzeppelin/contracts-upgradeable`, `@thirdweb-dev/contracts`, and `thirdweb` have been specified to exact versions. This update will make the build process more consistent and predictable by ensuring the same versions of dependencies are used each time. The `thirdweb:install` script is also removed as it's unnecessary with the fixed dependency versions.