Refactor shop item creation to item modification
This commit modifies the shop functions, changing the "_shopCreateItem" function into a "_shopModifyItem" function. The commit also removes the dependency on "NFTLogicDelegations.sol" and includes a new "_tokenBurnAndRedistribute" function to replace "_tokenBurnAndRedistribute" from the removed NFTLogicDelegations. Some slight modifications are performed on the "reinitializer_8_ShopLogic" and "ShopItemPurchased" to adapt to these changes.