Exclude burned plants from owner's plant information
In the getPlantInfo function, an adjustment was made to exclude burned plants from the returned details. This involved adding a conditional statement in the loop that checks, packs and returns array details on plants. The array that holds the information is then resized to match the number of valid plants (not burned).