Replace CustomRouter with PixotchiRouter in MainSetup2 tests
The test MainSetup2.sol file has been updated to replace the usage of the CustomRouter with the PixotchiRouter. This has led to the removal of the CustomRouter contract from the mentioned file. Also, the PixotchiRouter initialization now uses the default admin role as the message sender, rather than a hardcoded value.